+++ app/ui/src/pages/System/ObservabilityCard.tsx
@@ -5,7 +5,7 @@ import Box from '@mui/material/Box';
 import Chip from '@mui/material/Chip';
 import LinearProgress from '@mui/material/LinearProgress';
 import Typography from '@mui/material/Typography';
-import { fetchObservability } from '../../api/api';
+import { fetchObservability } from '../../api/systemAuditMetrics';
 import { queryKeys } from '../../api/queryKeys';
 import { SystemCardShell } from './SystemCardShell';
