+++ scripts/verify.local.env
@@ -1,0 +1,1 @@
+export BASE_URL="https://birdlense.eyera.info"
